implement user dataset for extension

+++ BackEnd/Effort.Models/Dto/TimesheetDto.cs
@@ -7,10 +7,10 @@ namespace Effort.Models.Dto
     public class TimesheetDto
     {
         public DateTime Date { get; set; }
-        public long ActivityTypeId { get; set; }
+        public int ActivityTypeId { get; set; }
         public String UserUniqueName { get; set; }
         public int WorkItemId { get; set; }
-        public long Duration { get; set; }
+        public int Duration { get; set; }
         public String Comment { get; set; }
     }
 }
